Move hooks into appropriate feature specific directories
