+++ frontend/src/features/noteTaking/common/hooks/useNotebookInfoQuery.hooks.ts
