+++ frontend/src/features/noteTaking/notes/hooks/useNotebookIdEffect.hooks.ts
